requirements.txt
azure-common==1.1.28
azure-core==1.26.4
azure-functions==1.15.0
azure-identity==1.12.0
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
langchain==0.0.121
matplotlib==3.7.1
numpy==1.23.5
openai==0.27.4
openpyxl==3.1.2
pandas==1.5.3
plotly==5.14.1
python-dotenv==1.0.0
pytest==7.3.1
scikit-learn==1.2.2
scipy==1.10.1
tiktoken==0.3.0
Unidecode==1.3.6
mongoengine==0.27.0
pymongo==4.3.3
Flask==2.2.2
Flask-Cors==3.0.10
bs4
werkzeug